+++ src/main/java/de/wbstraining/lotto/business/lottogesellschaft/AuftragGewinner.java
@@ -1,7 +1,6 @@
 package de.wbstraining.lotto.business.lottogesellschaft;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 import de.wbstraining.lotto.persistence.model.Gewinnklasseziehungquote;
@@ -23,8 +22,8 @@ public class AuftragGewinner implements Serializable {
 	private Lottoscheinziehung6aus49 lottoscheinziehung6aus49;
 
 	public AuftragGewinner(Lottoschein lottoschein, Kunde kunde, Ziehung ziehung,
-		Lottoscheinziehung lottoscheinziehung,
-		Lottoscheinziehung6aus49 lottoscheinziehung6aus49) {
+			Lottoscheinziehung lottoscheinziehung,
+			Lottoscheinziehung6aus49 lottoscheinziehung6aus49) {
 		this.lottoschein = lottoschein;
 		this.lottoscheinziehung = lottoscheinziehung;
 		this.kunde = kunde;
@@ -32,7 +31,7 @@ public class AuftragGewinner implements Serializable {
 		this.lottoscheinziehung6aus49 = lottoscheinziehung6aus49;
 	}
 
-	public BigInteger getBelegnummer() {
+	public Long getBelegnummer() {
 		return lottoschein.getBelegnummer();
 	}
 
@@ -58,7 +57,7 @@ public class AuftragGewinner implements Serializable {
 
 	public long getGewinnklsseid() {
 		return gewinnklasseziehungquote.getGewinnklasse()
-			.getGewinnklasseid();
+				.getGewinnklasseid();
 	}
 
 	public long getQuote() {
@@ -68,21 +67,21 @@ public class AuftragGewinner implements Serializable {
 	public long getGewinnspiel77() {
 		if (lottoscheinziehung.getGewinnspiel77() != null)
 			return lottoscheinziehung.getGewinnspiel77()
-				.longValue();
+					.longValue();
 		return 0;
 	}
 
 	public long getGewinnsuper6() {
 		if (lottoscheinziehung.getGewinnsuper6() != null)
 			return lottoscheinziehung.getGewinnsuper6()
-				.longValue();
+					.longValue();
 		return 0;
 	}
 
 	public long getGewinn() {
 		if (lottoscheinziehung6aus49 != null)
 			return lottoscheinziehung6aus49.getGewinn()
-				.longValue();
+					.longValue();
 		return 0;
 	}
